시험: #6972 회귀 시험을 src 단위 시험에서 통합 시험으로 옮긴다 (#6972)
Lint job 의 "Validate Rust test suite manifest" 가 source-side 단위 시험 증가를
막는다 — src/renderer/pagination.rs::tests#1 이 22 → 23, 총량이 4205 → 4206 이 됐다.

    [RustUnitTier] 오류: PR base 대비 source-side test 총량 증가 금지: 4206 > 4205
    PR base 대비 source unit test 증가 금지: src/renderer/pagination.rs::tests#1 (23 > 22)

시험을 tests/cases/ 통합 시험으로 옮겼다. find_inline_control_target_page 와
PageContent·ColumnContent·PageItem·PageLayoutInfo 가 모두 pub 이라 rhwp:: 경로로
그대로 이식된다. control_line_seg_index 는 pub(crate) 라 quirk 직접 단언은 뺐지만,
글자 위치 투영으로 되돌아가면 결과가 Some((0,0)) 이 아니라 None 이 되므로 단언 하나가
같은 것을 증명한다.

모호한 경우(같은 높이 줄 둘 = #2004 이미지 스택)에 종전 투영으로 되돌아가는지도
둘째 시험으로 함께 잠갔다.

검증: unit tier 4205(기준선 복귀), 매니페스트 --check 통과, 새 통합 시험 2/2 통과.

(cherry picked from commit 607e18a61a3ff9c06256b5ab05c441ecca74749f)
